+++ gatsby-node.js
@@ -1,6 +1,8 @@
 // Copyright (c) 2021 Antti Kivi
 // Licensed under the MIT License
 
-exports.createPages = require('./util/createPages');
-exports.onCreatePage = require('./util/onCreatePage');
-exports.onPostBuild = require('./util/onPostBuild');
+const createPages = require('./util/createPages');
+const onCreatePage = require('./util/onCreatePage');
+
+exports.createPages = createPages;
+exports.onCreatePage = onCreatePage;