+++ src/components/layout/Navigation.jsx
@@ -1,7 +1,8 @@
 // Copyright (c) 2021 Antti Kivi
 // Licensed under the MIT License
 
-import React, { useState } from 'react';
+import React from 'react';
+import PropTypes from 'prop-types';
 import { graphql, useStaticQuery } from 'gatsby';
 import styled, { css } from 'styled-components';
 
@@ -20,7 +21,7 @@ const Toggle = styled.div`
   cursor: pointer;
   user-select: none;
 
-  @media screen and ${(props) => props.theme.devices.tablet} {
+  @media screen and (${(props) => props.theme.devices.tablet}) {
     display: none;
   }
 `;
@@ -90,7 +91,7 @@ const Ul = styled.ul`
       max-height: none;
     `};
 
-  @media screen and ${(props) => props.theme.devices.tablet} {
+  @media screen and (${(props) => props.theme.devices.tablet}) {
     overflow: visible;
     display: flex;
     justify-content: center;
@@ -101,7 +102,7 @@ const Ul = styled.ul`
 const Li = styled.li`
   margin: 2rem 1em;
 
-  @media screen and ${(props) => props.theme.devices.tablet} {
+  @media screen and (${(props) => props.theme.devices.tablet}) {
     margin: 1em;
   }
 `;
@@ -129,71 +130,96 @@ const Link = styled(LocalizedLink)`
   }
 `;
 
-export default (props) => {
-  const data = useStaticQuery(
-    graphql`
-      query {
-        allContentfulMenu(filter: { contentful_id: { eq: "12OH6cgaTcp4TUDvpqslYc" } }) {
-          edges {
-            node {
-              node_locale
-              links {
-                ... on ContentfulIndexPage {
-                  contentful_id
-                  title
-                  internal {
-                    type
+const withMenuData = function withNavigationMenuQueryData(WrappedComponent) {
+  function WithMenuData(props) {
+    const data = useStaticQuery(
+      graphql`
+        query {
+          allContentfulMenu(filter: { contentful_id: { eq: "7oKEb5SnrTGF1vbDGwfBbr" } }) {
+            edges {
+              node {
+                node_locale
+                links {
+                  ... on ContentfulIndexPage {
+                    contentful_id
+                    title
+                    internal {
+                      type
+                    }
                   }
                 }
               }
             }
           }
         }
-      }
-    `,
-  );
+      `,
+    );
 
-  const [toggled, setToggled] = useState(false);
+    return <WrappedComponent data={data} {...props} />;
+  }
 
-  const handleClick = (event) => {
-    event.preventDefault();
-    setToggled(!toggled);
-  };
+  const wrappedComponentName = WrappedComponent.displayName || WrappedComponent.name || 'Component';
 
-  return (
-    <Nav>
-      <Toggle aria-controls="primary-menu" aria-expanded={toggled.toString()} onClick={handleClick}>
-        <ToggleBar toggled={toggled} />
-        <ToggleBar toggled={toggled} />
-        <ToggleBar toggled={toggled} />
-      </Toggle>
-      {/* TODO Consider getting the page titles and link values by query from Contentful */}
-      <Ul id="primary-menu" toggled={toggled}>
-        {data.allContentfulMenu.edges
-          .filter(({ node }) => node.node_locale === props.locale)[0]
-          .node.links.map((link) => {
-            switch (link.internal.type) {
-              case 'ContentfulIndexPage':
-                return (
-                  <Li key={link.contentful_id}>
-                    <Link to={link.contentful_id} locale={props.locale}>
-                      {link.title}
-                    </Link>
-                  </Li>
-                );
-              case 'ContentfulPath':
-                return (
-                  <Li key={link.contentful_id}>
-                    <Link to={link.contentful_id} locale={props.locale}>
-                      {link.title}
-                    </Link>
-                  </Li>
-                );
-              default:
-                return null;
-            }
-          })}
-      </Ul>
-    </Nav>
-  );
+  WithMenuData.displayName = `withMenuData(${wrappedComponentName})`;
+
+  return WithMenuData;
 };
+
+const propTypes = { data: PropTypes.object.isRequired, locale: PropTypes.string.isRequired };
+
+class Navigation extends React.Component {
+  constructor(props) {
+    super(props);
+
+    this.state = { isToggled: false };
+
+    this.handleClick = this.handleClick.bind(this);
+  }
+
+  handleClick(event) {
+    event.preventDefault();
+    this.setState((state, props) => ({ isToggled: !state.isToggled }));
+  }
+
+  render() {
+    const { data, locale } = this.props;
+
+    return (
+      <Nav>
+        <Toggle
+          aria-controls="primary-menu"
+          aria-expanded={this.state.isToggled.toString()}
+          onClick={this.handleClick}
+        >
+          <ToggleBar toggled={this.state.isToggled} />
+          <ToggleBar toggled={this.state.isToggled} />
+          <ToggleBar toggled={this.state.isToggled} />
+        </Toggle>
+        <Ul id="primary-menu" toggled={this.state.isToggled}>
+          {data.allContentfulMenu.edges
+            .filter(({ node }) => node.node_locale === locale)[0]
+            .node.links.map((link) => {
+              switch (link.internal.type) {
+                case 'ContentfulIndexPage': {
+                  return (
+                    <Li key={link.contentful_id}>
+                      <Link to={link.contentful_id} locale={locale}>
+                        {link.title}
+                      </Link>
+                    </Li>
+                  );
+                }
+                default: {
+                  return null;
+                }
+              }
+            })}
+        </Ul>
+      </Nav>
+    );
+  }
+}
+
+Navigation.propTypes = propTypes;
+
+export default withMenuData(Navigation);