+++ src/locales/en/footerMessages.js
@@ -2,12 +2,13 @@
 // Licensed under the MIT License
 
 export default {
+  footerLogoImageText: 'Antti’s signature',
   footerSocialMediaTitle: 'Social Media',
-  footerFacebookAlt: 'Facebook Logo',
-  footerGithubAlt: 'GitHub Logo',
-  footerInstagramAlt: 'Instagram Logo',
-  footerLinkedinAlt: 'LinkedIn Logo',
-  footerTwitterAlt: 'Twitter Logo',
+  footerFacebookImageText: 'Facebook Logo',
+  footerGithubImageText: 'GitHub Logo',
+  footerInstagramImageText: 'Instagram Logo',
+  footerLinkedinImageText: 'LinkedIn Logo',
+  footerTwitterImageText: 'Twitter Logo',
   footerOcticons: 'Octicons&nbsp;&copy;&nbsp;2021&nbsp;GitHub,&nbsp;Inc.',
   footerCopyright: '&copy;&nbsp;2021&nbsp;Antti&nbsp;Kivi',
   footerMadeBy: 'The site is rendered through the lens of <a>Visiosto</a>',