+++ src/templates/NotFound.jsx
@@ -1,0 +1,87 @@
+// Copyright (c) 2021 Antti Kivi
+// Licensed under the MIT License
+
+import React from 'react';
+import PropTypes from 'prop-types';
+import { graphql } from 'gatsby';
+import styled from 'styled-components';
+import { useIntl } from 'react-intl';
+
+import Intl from '../components/Intl';
+import LayoutError from '../components/layout/LayoutError';
+import Theme from '../components/Theme';
+
+import createInternationalization from '../util/createInternationalization';
+
+const Div = styled.div`
+  margin: 2em ${(props) => props.theme.layout.marginMobile};
+  text-align: center;
+
+  @media screen and (${(props) => props.theme.devices.mobileL}) {
+    margin: 2em ${(props) => props.theme.layout.marginTablet};
+  }
+
+  @media screen and (${(props) => props.theme.devices.tablet}) {
+    margin: 2em ${(props) => props.theme.layout.marginDesktop};
+  }
+`;
+
+const propTypes = {
+  children: PropTypes.node,
+  data: PropTypes.object.isRequired,
+  location: PropTypes.object.isRequired,
+  navigate: PropTypes.func.isRequired,
+  pageContext: PropTypes.object.isRequired,
+  pageResources: PropTypes.object.isRequired,
+  params: PropTypes.object.isRequired,
+  path: PropTypes.string.isRequired,
+  uri: PropTypes.string.isRequired,
+};
+
+const defaultProps = { children: undefined };
+
+function Page({ pageContext }) {
+  const intl = createInternationalization(useIntl());
+  const { locale, pageID } = pageContext;
+
+  return (
+    <LayoutError errorCode="404" locale={locale} pageID={pageID} title={intl('notFoundTitle')}>
+      <Div>
+        <p>{intl('notFoundContent')}</p>
+      </Div>
+    </LayoutError>
+  );
+}
+
+Page.propTypes = propTypes;
+Page.defaultProps = defaultProps;
+
+function NotFound(props) {
+  const { simpleLocales } = props.data.site.siteMetadata;
+  const { locale } = props.pageContext;
+  return (
+    <Intl locale={simpleLocales[locale.replace('-', '_')]}>
+      <Theme>
+        <Page {...props} />
+      </Theme>
+    </Intl>
+  );
+}
+
+NotFound.propTypes = propTypes;
+NotFound.defaultProps = defaultProps;
+
+export default NotFound;
+
+export const pageQuery = graphql`
+  query NotFoundQuery {
+    site {
+      siteMetadata {
+        simpleLocales {
+          en_GB
+          fi
+        }
+      }
+    }
+  }
+`;